+++ plugins/scaffolder-actions/scaffolder-backend-module-utils/src/actions/jsonata/jsonata.test.ts
@@ -15,14 +15,14 @@
  */
 import { PassThrough } from 'stream';
 import { createJSONataAction } from './jsonata';
-import { getVoidLogger } from '@backstage/backend-common';
+import { mockServices } from '@backstage/backend-test-utils';
 
 describe('roadiehq:utils:jsonata', () => {
   const mockContext = {
     task: {
       id: 'task-id',
     },
-    logger: getVoidLogger(),
+    logger: mockServices.logger.mock(),
     logStream: new PassThrough(),
     output: jest.fn(),
     createTemporaryDirectory: jest.fn(),