+++ plugins/backend/backstage-aws-backend/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-aws-backend
 
+## 1.5.0
+
+### Minor Changes
+
+- c2274f9: Upgrade backstage version to `1.44.2`.
+
 ## 1.4.0
 
 ### Minor Changes